conecte el Home con la api, limpie el codigo de comentarios, reestructure constantes y la pase a un objeto Query_sql, movi las funciones de la API q no estan en uso en archivo funcionesParaProyecto.py, movi logica que serializa el producto a su archivo correspondiente serializer, cambie el nombre de api.py por app.py, la ejecucion de la query la va tener el objeto Query_sql con el metodo ejecutarSQL
